# GVT-3232 Optimize publication log / ext-api / geocoding