--- a/infra/src/main/kotlin/fi/fta/geoviite/infra/publication/PublicationLogRemarks.kt
+++ b/infra/src/main/kotlin/fi/fta/geoviite/infra/publication/PublicationLogRemarks.kt
@@ -3,12 +3,11 @@ package fi.fta.geoviite.infra.publication
 import fi.fta.geoviite.infra.common.KmNumber
 import fi.fta.geoviite.infra.common.TrackMeter
 import fi.fta.geoviite.infra.common.TrackNumber
-import fi.fta.geoviite.infra.geography.calculateDistance
 import fi.fta.geoviite.infra.localization.Translation
 import fi.fta.geoviite.infra.localization.localizationParams
 import fi.fta.geoviite.infra.math.Point
+import fi.fta.geoviite.infra.math.lineLength
 import fi.fta.geoviite.infra.math.roundTo1Decimal
-import fi.fta.geoviite.infra.tracklayout.LAYOUT_SRID
 import kotlin.math.abs
 
 fun publicationChangeRemark(translation: Translation, key: String, value: String? = null) =
@@ -50,7 +49,9 @@ fun getPointMovedRemarkOrNull(
     return oldPoint?.let { p1 ->
         newPoint?.let { p2 ->
             if (!pointsAreSame(p1, p2)) {
-                val distance = calculateDistance(listOf(p1, p2), LAYOUT_SRID)
+                // TODO: GVT-3232 How useful is this?
+                val distance = lineLength(p1, p2)
+                //                val distance = calculateDistance(listOf(p1, p2), LAYOUT_SRID)
                 if (distance > DISTANCE_CHANGE_THRESHOLD) {
                     publicationChangeRemark(translation, translationKey, formatDistance(distance))
                 } else {