Nippu optimointeja julkaisluokille ja ext-api:lle ja vähän muuallekin missä vain tehdään reverse geokoodausta, lokalisaatiota, ja historiallisen kaman käsittelyä. Yksittäin jokainen näistä on hyvinkin pieniä asioita, mutta kokonaisuutena niistä kertyi ihan kohtuullinen määrä. 

Testailin ennenkaikkea tietyllä pitkällä julkaisulokin välillä, jossa näiden kokonaisvaikutus oli hiukan yli puolet:
- Ensimmäinen haku (paljon asioita haettava cacheen): 7.5s -> 3.5s
- Myöhemmät identtiset haut (lähinnä asioiden uudelleenlaskentaa ja julkaisujen sisältöjen haut): 4s -> 1.5s
Tältä pohjalta voitaisiin harkita kasvattaa julkaisulokin maksimiväli 6kk -> 12kk. Yksi vuosi tuntuisi intuitiivisemmalta rajalta kuin puoli vuotta.

Muuttuneita tiedostoja on iso kasa johtuen tiettyjen yleisten funkkareiden käyttötavan muutoksesta, mutta ne on nopea plarata läpi koska identtisiä. Pointtaan ne oleelliset paikat kommenteilla.